% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaccaroZaki.R
\name{acov}
\alias{acov}
\title{Empirical Allan covariance}
\usage{
acov(X)
}
\arguments{
\item{Xt}{A \code{matrix} of dimension T by p, where T is the length of the time series and p is the number of processes.}
}
\value{
A \code{list} with the following structure:
\itemize{
 \item covariance: A \code{list} of \code{matrix} for each level \eqn{m} of the estimated Allan covariance
 \item scales: A \code{vector} of the employed levels (\eqn{m} in [1])
}
}
\description{
This function provides an empirical estimate of the Allan covariance given multiple processes.
The detailed definition can be found in Equation (23) in Richard J. Vaccaro and Ahmed S. Zaki, "Reduced-Drift Virtual Gyro from an Array of Low-Cost Gyros".
}
\author{
Davide Antonio Cucci
}
