% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaccaroZaki.R
\name{av}
\alias{av}
\title{Empirical Allan variance}
\usage{
av(Xt)
}
\arguments{
\item{Xt}{A \code{matrix} of dimension T by p, where T is the length of the time series and p is the number of processes.}
}
\value{
A \code{list} with the following structure:
\itemize{
 \item J TO COMPLETE
 \item scales TO COMPLETE
 \item variance TO COMPLETE
}
}
\description{
This function provides an empirical estimate of the Allan variance given multiple processes.
The detailed definition can be found in Equation (3) in Richard J. Vaccaro and Ahmed S. Zaki, "Reduced-Drift Virtual Gyro from an Array of Low-Cost Gyros".
}
\author{
Davide Antonio Cucci
}
