% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BMI055}
\alias{BMI055}
\title{BMI055 Gyroscope Data}
\format{
A \code{matrix} with 4194305 rows and 6 columns corresponding to each of the collected sequence.
}
\usage{
BMI055
}
\description{
Six static data sequences collected at 200 Hz for the Bosch BMI055 gyroscope.
}
\keyword{datasets}
