% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{d.wccv_RW}
\alias{d.wccv_RW}
\title{TO DO}
\usage{
d.wccv_RW(J)
}
\arguments{
\item{J}{what it does}

\item{R}{what it does}
}
\value{
something
}
\description{
WHAT IT DOES
}
\details{
If needed
}
\examples{
# TO BE ADDED

}
\author{
WHO WROTE IT
}
