% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{est_cov}
\alias{est_cov}
\title{Empirical Covariance of Coefficients on Individual Signals}
\usage{
est_cov(Xt, scale_weights, sB = 10, B)
}
\arguments{
\item{Xt}{A \code{matrix} of dimension T by p, where T is the length of the time series and p is the number of processes.}

\item{scale_weights}{A \code{vector} that denotes the weights on scales. All elements should be non-negative and sum to one.}

\item{sB}{A \code{double} that denotes the positive constant C associated with the block size, which is defined as floor(C*T^{1/3}). Default value is 10.}

\item{B}{An \code{integer} indicating the number of Monte-Carlo replications used in the Moving Block Bootstrap.}
}
\value{
A \code{matrix} of the estimated covariance matrix of the coefficients on individual signals.
}
\description{
This function computes the estimated covariance matrix of the coefficients on individual signals using the Moving Block Bootstrap approach considered in Zhang et al. (2021).
The detailed definition can be found in Equation (8) in Zhang et al. (2021) (https://arxiv.org/abs/2106.15997).
}
\author{
Yuming Zhang
}
