% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation.R
\name{expect.wv_WN}
\alias{expect.wv_WN}
\title{TO DO}
\usage{
expect.wv_WN(R, J)
}
\arguments{
\item{R}{what it does}

\item{J}{what it does}
}
\value{
something
}
\description{
WHAT IT DOES
}
\details{
If needed
}
\examples{
# TO BE ADDED
}
\author{
WHO WROTE IT
}
