% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_optimal_coefs_iterative}
\alias{find_optimal_coefs_iterative}
\title{iterative method for optimal coefs}
\usage{
find_optimal_coefs_iterative(Xt, scale_weights)
}
\description{
iterative method for optimal coefs
}
