% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_optimal_coefs_theo}
\alias{find_optimal_coefs_theo}
\title{closed-form solution to optimal coefs from theoretical wvs and wccvs}
\usage{
find_optimal_coefs_theo(wccv.mat, wv.mat, scale_weights)
}
\description{
closed-form solution to optimal coefs from theoretical wvs and wccvs
}
