% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaccaroZaki.R
\name{find_optimal_coefs_vaccaro}
\alias{find_optimal_coefs_vaccaro}
\title{Estimate Optimal Coefficients based on Reduced-Drift Virtual Gyro}
\usage{
find_optimal_coefs_vaccaro(Q)
}
\arguments{
\item{Q}{\code{matrix} of size \eqn{p} by \eqn{p} specifying the covariance of the Random Walk innovation.}
}
\value{
c \code{vector} of coefficients.
}
\description{
This function computes the optimal coefficients for a linear combination of \eqn{p} gyroscopes according to
Equation (43) in Richard J. Vaccaro and Ahmed S. Zaki, "Reduced-Drift Virtual Gyro from an Array of Low-Cost Gyros".
}
\author{
Davide Antonio Cucci
}
