% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaccaroZaki.R
\name{find_optimal_coefs_vaccaro_pinv}
\alias{find_optimal_coefs_vaccaro_pinv}
\title{Estimate Optimal Coefficients based on Reduced-Drift Virtual Gyro}
\usage{
find_optimal_coefs_vaccaro_pinv(Q, n)
}
\arguments{
\item{Q}{A \code{matrix} of size \eqn{p} by \eqn{p} specifying the covariance of the random walk innovation.}

\item{n}{An \code{integer} indicating the number of singular values to ignore in the calculation of the pseudo-inverse of \eqn{Q}.}
}
\value{
c A \code{vector} of coefficients.
}
\description{
This function computes the optimal coefficients for a linear combination of \eqn{p} gyroscopes according to
Equation (43) in Richard J. Vaccaro and Ahmed S. Zaki, "Reduced-Drift Virtual Gyro from an Array of Low-Cost Gyros", where
\eqn{Q^{-1}} is replaced by the pseudo-inverse of \eqn{Q}, ignoring \eqn{n} singular values, as suggested in Equation (54).
}
\author{
Davide Antonio Cucci
}
