% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_c_var}
\alias{get_c_var}
\title{get estimated covariance matrix for the coefficients with block bootstrap}
\usage{
get_c_var(Xt, scale_weights, c_hat, B = 10^3, sB = 10)
}
\description{
get estimated covariance matrix for the coefficients with block bootstrap
}
