% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_virtual_gyro}
\alias{get_virtual_gyro}
\title{Construct Virtual Gyroscope Signal}
\usage{
get_virtual_gyro(Xt, weights)
}
\arguments{
\item{Xt}{A \code{matrix} of dimension T by p, where T is the length of the time series and p is the number of processes.}

\item{weights}{A \code{vector} of coefficients on individual signals. All elements should be non-negative and sum to one.}
}
\value{
A \code{vector} of the virtual gyroscope signal.
}
\description{
This function computes the virtual gyroscope signal by taking a linear combination of the individual gyroscope signals.
}
\author{
Yuming Zhang
}
