% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ICM20689}
\alias{ICM20689}
\title{ICM20689 Gyroscope Data}
\format{
A \code{matrix} with 4194305 rows and 6 columns corresponding to each of the collected sequence.
}
\usage{
ICM20689
}
\description{
Six static data sequences collected at 200 Hz for the InvenSense ICM-20689 gyroscope.
}
\keyword{datasets}
