% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_virtual_gyro}
\alias{plot_virtual_gyro}
\title{Wavelet Variance Plot for Virtual Gyroscopes}
\usage{
plot_virtual_gyro(Xt, ..., names, ratio_ylim = NA)
}
\arguments{
\item{Xt}{A \code{matrix} of dimension T by p, where T is the length of the time series and p is the number of processes.}

\item{...}{Multiple \code{vector} of coefficients.}

\item{names}{A \code{vector} of strings to be used for the legend.}

\item{ration_ylim}{Custom y limits for the ratio plot}
}
\description{
This function plots the wavelet variance of the obtained virtual gyroscopes with the provided coefficients
}
\author{
Davide Antonio Cucci
}
