% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_corr_ar1}
\alias{simulate_corr_ar1}
\title{Simulate Correlated AR(1)}
\usage{
simulate_corr_ar1(N, phi_ar1, sigma_ar1)
}
\arguments{
\item{N}{An \code{integer} indicating the length of the time series.}

\item{phi_ar1}{A \code{vector} indicating the phi terms of the AR(1) processes.}

\item{sigma_ar1}{A \code{matrix} denoting the covariance of the AR(1) innovations.}
}
\value{
A \code{matrix} of the simulated correlated AR(1) processes.
}
\description{
This function simulates multiple correlated AR(1) processes.
}
\author{
Davide Antonio Cucci
}
