% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_corr_rw}
\alias{simulate_corr_rw}
\title{Simulate Correlated Random Walk}
\usage{
simulate_corr_rw(N, sigma_rw)
}
\arguments{
\item{N}{An \code{integer} indicating the length of the time series.}

\item{sigma_rw}{A \code{matrix} denoting the covariance of the random walk innovations.}
}
\value{
A \code{matrix} of the simulated correlated random walk processes.
}
\description{
This function simulates multiple correlated random walk processes.
}
\author{
Davide Antonio Cucci
}
