% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_wn}
\alias{simulate_wn}
\title{Simulate White Noise}
\usage{
simulate_wn(N, sigma_wn)
}
\arguments{
\item{N}{An \code{integer} indicating the length of the time series.}

\item{sigma_wn}{A \code{vector} of variances of the white noise processes.}
}
\value{
A \code{matrix} of the simulated white noise processes.
}
\description{
This function simulates multiple white noise processes.
}
\author{
Davide Antonio Cucci
}
