% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_wn_corr_ar1_corr_rw}
\alias{simulate_wn_corr_ar1_corr_rw}
\title{Simulates a WN + correlated AR1 + coorelated RW}
\usage{
simulate_wn_corr_ar1_corr_rw(N, sigma_wn, phi_ar1, sigma_ar1, sigma_rw)
}
\arguments{
\item{N}{the number of samples}

\item{sigma_wn}{a vector of size \code{n} of WN variances for each time series}

\item{phi_ar1}{a vector of size \code{n}, the phi of the AR1s}

\item{sigma_ar1}{a matrix of size \code{n}x\code{n}, the variance of AR1s innovations}

\item{sigma_rw}{a matrix of size \code{n}x\code{n}, the covariance of RW innovations}
}
\description{
Simulates a WN + correlated AR1 + coorelated RW
}
