% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simulate_wn_corr_rw_ts}
\alias{simulate_wn_corr_rw_ts}
\title{Simulates a WN + coorelated RW from a T-Student distribution}
\usage{
simulate_wn_corr_rw_ts(N, sigma_wn, sigma_rw, df = 5)
}
\arguments{
\item{N}{the number of samples}

\item{sigma_wn}{a vector of size \code{n} of WN variances for each time series}

\item{sigma_rw}{a matrix of size \code{n}x\code{n}, the covariance of RW innovations}
}
\description{
Simulates a WN + coorelated RW from a T-Student distribution
}
