% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{theo_wv_wn_corr_ar1_corr_rw}
\alias{theo_wv_wn_corr_ar1_corr_rw}
\title{Theoretical wavelet variance of a WN + correlated AR1 + coorelated RW}
\usage{
theo_wv_wn_corr_ar1_corr_rw(sigma_wn, phi_ar1, sigma_ar1, sigma_rw, J)
}
\arguments{
\item{sigma_wn}{a vector of size \code{n} of WN variances for each time series}

\item{phi_ar1}{a vector of size \code{n}, the phi of the AR1s}

\item{sigma_ar1}{a matrix of size \code{n}x\code{n}, the variance of AR1s innovations}

\item{sigma_rw}{a matrix of size \code{n}x\code{n}, the covariance of RW innovations}

\item{J}{maximum scale, so that tau = 2^(1:J)}
}
\description{
Theoretical wavelet variance of a WN + correlated AR1 + coorelated RW
}
