% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{theo_wv_wn_corr_rw}
\alias{theo_wv_wn_corr_rw}
\title{Theoretical wavelet variance of a WN  + coorelated RW}
\usage{
theo_wv_wn_corr_rw(sigma_wn, sigma_rw, J)
}
\arguments{
\item{sigma_wn}{a vector of size \code{n} of WN variances for each time series}

\item{sigma_rw}{a matrix of size \code{n}x\code{n}, the covariance of RW innovations}

\item{J}{maximum scale, so that tau = 2^(1:J)}
}
\description{
Theoretical wavelet variance of a WN  + coorelated RW
}
