% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{wccv_local}
\alias{wccv_local}
\title{Empirical Wavelet Cross-Covariance}
\usage{
wccv_local(Xt)
}
\arguments{
\item{Xt}{A \code{matrix} of dimension T by p, where T is the length of the time series and p is the number of processes.}
}
\value{
A \code{list} with the following structure:
\itemize{
 \item wccv: A \code{matrix} of the estimated wavelet cross-covariance.
 \item wccv.cov: A \code{matrix} of the estimated covariance matrix of the estimated wavelet cross-covariance.
 \item ci_low: A \code{matrix} of the lower bound of the confidence interval of the estimated wavelet cross-covariance.
 \item ci_high: A \code{matrix} of the upper bound of the confidence interval of the estimated wavelet cross-covariance.
}
}
\description{
This function provides an empirical estimate of the wavelet cross-covariance given multiple processes.
}
\author{
Haotian Xu
}
