% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VaccaroZaki.R
\name{av}
\alias{av}
\title{Empirical Allan variance}
\usage{
av(Xt)
}
\arguments{
\item{Xt}{A \code{matrix} of dimension T by p, where T is the length of the time series and p is the number of processes.}
}
\value{
A \code{list} with the following structure:
\itemize{
 \item J: An \code{integer} indicating the maximum level.
 \item scales: A \code{vector} of the employed levels (\eqn{N} in [1]).
 \item variance: A \code{matrix} of the estimated Allan variance.
}
}
\description{
This function provides an empirical estimate of the Allan variance given multiple processes.
The detailed definition can be found in Equation (3) in Richard J. Vaccaro and Ahmed S. Zaki, "Reduced-Drift Virtual Gyro from an Array of Low-Cost Gyros".
}
\author{
Davide Antonio Cucci
}
